requirements.txt
torch
torchvision
matplotlib
scikit-learn
numpy
Pillow